envoie kasifinal

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: droly <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/26 16:13:31 by droly             #+#    #+#             */
-/*   Updated: 2017/04/26 13:50:48 by droly            ###   ########.fr       */
+/*   Updated: 2017/05/10 15:12:52 by droly            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void		*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*str;
 	char	*src1;
 
 	i = 0;
 	str = (char*)dst;
 	src1 = (char*)src;
-	while (i < (int)n)
+	while (i < n)
 	{
 		str[i] = src1[i];
 		i++;
